Move toggle button out of history menu
